[test](iceberg) Isolate async admission policy
